move feature file to ext2020
