+++ src/events/forms.py
@@ -2,7 +2,6 @@ from django import forms
 
 from .models import CustomEvent, Schedule
 
-from ext2020.models import Venue
 
 
 class ScheduleCreationForm(forms.ModelForm):
@@ -23,8 +22,3 @@ class CustomEventForm(forms.ModelForm):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.fields['title'].strip = False
-
-class CommunityTrackForm(forms.Form):
-    venue = forms.ModelChoiceField(queryset=Venue.objects.all())
-    attendee_token = forms.CharField()
-    selected_time = forms.DateTimeField()
