+++ src/events/views.py
@@ -15,7 +15,7 @@ from core.mixins import FormValidMessageMixin
 from core.utils import OrderedDefaultDict, TemplateExistanceStatusResponse
 from proposals.models import AdditionalSpeaker, TalkProposal, TutorialProposal
 
-from .forms import ScheduleCreationForm, CommunityTrackForm
+from .forms import ScheduleCreationForm
 from .models import (
     EVENT_ROOMS, Schedule, Time,
     CustomEvent, KeynoteEvent, SponsoredEvent,
@@ -24,10 +24,7 @@ from .models import (
 from .renderers import render_all
 
 
-from core.utils import collect_language_codes
-import datetime
-from django.core.exceptions import ImproperlyConfigured
-from ext2020.models import Venue, Choice
+
 logger = logging.getLogger(__name__)
 
 
@@ -291,48 +288,3 @@ class TutorialDetailView(
     event_model = ProposedTutorialEvent
     template_name = 'events/tutorial_detail.html'
     response_class = TemplateExistanceStatusResponse
-
-
-class CommunityTrackView(ListView):
-    model = Venue
-    path = 'events/community-track'
-    success_url = reverse_lazy('community-track')
-
-    def get(self, request, *args, **kwargs):
-        return super().get(request, *args, **kwargs)
-
-    def post(self, request, *args, **kwargs):
-
-        data = request.POST.copy()
-        data['selected_time'] = datetime.datetime.now()
-        forms_obj = CommunityTrackForm(data)
-        if forms_obj.is_valid():
-            Choice.objects.create(**forms_obj.cleaned_data)
-            return HttpResponseRedirect(self.get_success_url())
-
-
-
-
-    def get_context_data(self, **kwargs):
-        token = self.request.GET.get('token')
-        venue_choice = Venue.objects.filter(choice__attendee_token=token)
-        kwargs.update({
-            'venue_choice': venue_choice.first() if venue_choice else '',
-            'token': token,
-        })
-        return super().get_context_data(**kwargs)
-
-    def get_template_names(self):
-        template_names = [
-            '/'.join(['contents', code, self.path + '.html'])
-            for code in collect_language_codes(self.request.LANGUAGE_CODE)
-        ]
-        # print(template_names)
-        return template_names
-
-    def get_success_url(self):
-        """Return the URL to redirect to after processing a valid form."""
-        if not self.success_url:
-            raise ImproperlyConfigured("No URL to redirect to. Provide a success_url.")
-        return str(self.success_url)  # success_url may be lazy
-
